pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cerebras-cloud-sdk>=1.67.0",
    "load-dotenv>=0.1.0",
    "openai>=2.24.0",
    "python-json-logger>=4.0.0",
    "setuptools>=82.0.0",
    "streamlit>=1.54.0",
]
